seems sqlite wants these to be longs

+++ model/Character.cs
@@ -9,11 +9,11 @@ namespace evefifo.model
     public class Character
     {
         public string CloneName { get; set; }
-        public int CloneSP { get; set; }
+        public long CloneSP { get; set; }
         public string CorpName { get; set; }
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public double SecStatus { get; set; }
-        public int SP { get; set; }
+        public long SP { get; set; }
     }
 }
